+++ BALSAMIC/commands/report/deliver.py
@@ -9,15 +9,13 @@ import datetime
 import subprocess
 from pathlib import Path
 
-from BALSAMIC.constants.quality_check_reporting import METRICS_TO_DELIVER
-from BALSAMIC.utils.cli import get_file_extension
+from BALSAMIC.utils.cli import get_file_extension, read_yaml
 from BALSAMIC.utils.cli import write_json
 from BALSAMIC.utils.cli import get_snakefile
 from BALSAMIC.utils.cli import SnakeMake
 from BALSAMIC.utils.cli import convert_deliverables_tags
-from BALSAMIC.utils.rule import get_result_dir, get_capture_kit
+from BALSAMIC.utils.rule import get_result_dir
 from BALSAMIC.utils.exc import BalsamicError
-from BALSAMIC.utils.qc_metrics import get_qc_metrics_json, extract_metrics_for_delivery
 from BALSAMIC.utils.qc_report import render_html, report_data_population
 from BALSAMIC.constants.workflow_params import VCF_DICT
 from BALSAMIC.constants.workflow_rules import DELIVERY_RULES
@@ -86,14 +84,6 @@ LOG = logging.getLogger(__name__)
     help=f"Run workflow with selected variant caller(s) disable. Use comma to remove multiple variant callers. Valid "
     f"values are: {list(VCF_DICT.keys())}",
 )
-@click.option(
-    "--qc-metrics/--no-qc-metrics",
-    default=True,
-    show_default=True,
-    is_flag=True,
-    help=f"Generates a YAML file of quality control metrics. "
-    f"Currently retrieved metrics: {', '.join(list(set(METRICS_TO_DELIVER['targeted'] + METRICS_TO_DELIVER['wgs'])))}",
-)
 @click.pass_context
 def deliver(
     context,
@@ -104,7 +94,6 @@ def deliver(
     disable_variant_caller,
     sample_id_map,
     case_id_map,
-    qc_metrics,
 ):
     """
     cli for deliver sub-command.
@@ -163,10 +152,13 @@ def deliver(
         meta["case_name"] = case_id_map[1]
         meta["apptag"] = case_id_map[2]
 
-        collected_qc = get_qc_metrics_json(
-            sample_config_dict["analysis"]["result"],
-            sequencing_type,
-            get_capture_kit(sample_config_dict),
+        collected_qc = read_yaml(
+            os.path.join(
+                sample_config_dict["analysis"]["result"],
+                "qc",
+                sample_config_dict["analysis"]["case_id"]
+                + "_metrics_deliverables.yaml",
+            )
         )
         meta = report_data_population(collected_qc=collected_qc, meta=meta)
         balsamic_qc_report = os.path.join(
@@ -262,30 +254,6 @@ def deliver(
                 "step": "balsamic_delivery",
                 "format": get_file_extension(balsamic_qc_report),
                 "tag": ["coverage-qc-report"],
-                "id": case_name,
-            }
-        )
-
-    # Add output metrics delivery to report
-    if qc_metrics:
-        metric_delivery_report = os.path.join(
-            yaml_write_directory, case_name + "_metrics_deliverables.yaml"
-        )
-        metrics = extract_metrics_for_delivery(
-            sample_config_dict["analysis"]["result"], sequencing_type
-        )
-
-        with open(metric_delivery_report, "w") as fn:
-            yaml.dump(metrics, fn, default_flow_style=False)
-
-        LOG.info(f"Created metrics delivery file: {metric_delivery_report}")
-
-        delivery_json["files"].append(
-            {
-                "path": metric_delivery_report,
-                "step": "balsamic_delivery",
-                "format": get_file_extension(metric_delivery_report),
-                "tag": ["qc-metrics-yaml"],
                 "id": case_name,
             }
         )