+++ tests/commands/report/test_deliver.py
@@ -35,6 +35,8 @@ def test_deliver_tumor_only_panel(
                 "tumor:tumor:KS454",
                 "--case-id-map",
                 "gmck-solid:KSK899:apptag",
+                "--disable-variant-caller",
+                "cnvkit",
             ]
         )
 
@@ -107,43 +109,3 @@ def test_deliver_tumor_normal_panel(
         assert result.exit_code == 0
         assert actual_delivery_report.is_file()
         assert "following" in caplog.text
-
-
-def test_deliver_metrics(
-    invoke_cli,
-    environ,
-    tumor_normal_config,
-    helpers,
-    sentieon_install_dir,
-    sentieon_license,
-    caplog,
-):
-
-    # GIVEN a tumor-normal config file
-    helpers.read_config(tumor_normal_config)
-    actual_metric_delivery_yaml = Path(
-        helpers.delivery_dir, helpers.case_id + "_metrics_deliverables.yaml"
-    )
-
-    with mock.patch.dict(
-        environ,
-        {
-            "SENTIEON_LICENSE": sentieon_license,
-            "SENTIEON_INSTALL_DIR": sentieon_install_dir,
-        },
-    ), caplog.at_level(logging.DEBUG):
-        # WHEN running analysis
-        result = invoke_cli(
-            [
-                "report",
-                "deliver",
-                "--sample-config",
-                tumor_normal_config,
-                "--qc-metrics",
-            ]
-        )
-
-    # THEN it should run without any error
-    assert result.exit_code == 0
-    assert actual_metric_delivery_yaml.is_file()
-    assert "following" in caplog.text